`static_cast<GLfloat>` before storing into `vertex.y`.
All tests pass.

	modified:   code/ylikuutio/geometry/triangulation_templates.hpp

+++ code/ylikuutio/geometry/triangulation_templates.hpp
@@ -90,13 +90,13 @@ namespace geometry
                     uint32_t current_vertex_i = image_width * z + x;
 
                     // Interpolate y coordinate (altitude).
-                    GLfloat y = center_y(x, z, input_vertex_pointer, image_width);
+                    float y = center_y(x, z, input_vertex_pointer, image_width);
 
                     // Create a new vertex using bilinear interpolation.
                     // This corresponds to "v": specify one vertex.
                     glm::vec3 vertex;
                     vertex.x = static_cast<GLfloat>(x) - 0.5f;
-                    vertex.y = y;
+                    vertex.y = static_cast<GLfloat>(y);
                     vertex.z = static_cast<GLfloat>(z) - 0.5f;
                     temp_vertices.push_back(vertex);
